feat(settings): add settings screen navigation from app menu
