+++ package.json
@@ -109,7 +109,7 @@
     "react": "^18.2.0",
     "react-audio-visualize": "^1.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.16.0"
+    "react-router": "^7.4.0"
   },
   "devDependencies": {
     "@electron/rebuild": "^3.3.0",
