+++ src/renderer/screens/index.ts
@@ -1,0 +1,2 @@
+export * from './Home';
+export * from './Settings';
