+++ src/types.ts
@@ -1,0 +1,8 @@
+import { RENDERER_ROUTE, RENDERER_ROUTE_ACTION } from './constants';
+
+type UpdateRendererRouteArgs = {
+  action: RENDERER_ROUTE_ACTION;
+  route: RENDERER_ROUTE;
+};
+
+export type { UpdateRendererRouteArgs };
